pyproject.toml
[tool.poetry]
name = "dio-banking-api"
version = "0.1.0"
description = "DIO Async API challenge"
authors = [" guilherme-biscaino  < guilhermebs.bs2@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
sqlalchemy = "^2.0.36"
pydantic = "^2.10.1"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.1"
pyjwt = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
